// src/app/modules/admissionSemester/admissionSemester.interface.ts
export type TMonths =
    | "January"
    | "February"
    | "March"
    | "April"
    | "May"
    | "June"
    | "July"
    | "August"
    | "September"
    | "October"
    | "November"
    | "December";
export type TAcademicSemesterName = "Autumn" | "Summer" | "Fall";
export type TAcademicSemesterCode = "01" | "02" | "03";
export type TAdmissionSemester = {
    name: TAcademicSemesterName;
    year: Date;
    code: TAcademicSemesterCode;
    startMonth: TMonths;
    endMonth: TMonths;
};